Menu and Gargoyle and Spawner
I created a placeholder for our Main Menu, and drew out the first
background picture, with a Skybox in the main camera. Next I imported
all the gargoyle sprites and made a prefab, I created one idle
animation. I made a new EnemySpawner to work with the infinite runner
concept. I added a circle collider to the player and adjusted its
animations in order to allow the player to be not kinematic. Z rotation
is now restricted. I made a skeleton with shield prefab.