Drawbridge, Skeleton, OnCollisionIdle,
Drawbridge: Added a collider so arrows can’t go through it that turns
off once opened. Also changed script to rely on the OnCollisionIdle
script

OnCollisionIdle: sets the boolean run state to false of anything with
an animator component that collides with it. Thus, having their
animation state go to idle. A public boolean is available to set from
outside the script to make it so they can move again

Skeleton: Made a new skeletonIdle animation (problems with sidle), and
updated the animator component. Also SkeletonAI was adjusted to have
the skeleton stop moving when in the idle state.