Changed Thrust to Drift, fixed UI
The UI was varying per screen resolution, Fixated that. Additionally I
locked the Aspect Ratio to 16:9, which is the typical widescreen tv.
Finally, I changed it so that the trigger now acts as drift instead of
thrust, didn’t bother changing the variable names, but its much more
intuitive now for the user.