Main menu any time
Also added a main menu option to controls (never need to but can use B
button to go back), and updated those pictures to show boost. The start
button will now overlay the options menu on top of character selection,
game mode selection, and stage selection, with the additional options
to resume or to go back to the main menu.